+++ src/TestLogger/Core/TestCaseNameParser.cs
@@ -82,21 +82,5 @@ namespace Spekt.TestLogger.Core
 
             return pn;
         }
-
-        public class ParsedName
-        {
-            public ParsedName(string namespaceName, string typeName, string methodName)
-            {
-                this.NamespaceName = namespaceName;
-                this.TypeName = typeName;
-                this.MethodName = methodName;
-            }
-
-            public string NamespaceName { get; }
-
-            public string TypeName { get; }
-
-            public string MethodName { get; }
-        }
     }
 }
